Move TokenSigner to a separate file
Move TokenSigner to it’s own file for better readability
