+++ src/index.ts
@@ -1,8 +1,8 @@
 export { FederatedGraphQLDataSource } from "./datasource";
 export { GatewayAuthPlugin } from "./gateway";
-export { PublicFederatedToken, TokenSigner } from "./jwt";
+export { PublicFederatedToken } from "./jwt";
 export { FederatedAuthPlugin } from "./plugin";
+export { TokenSigner } from "./sign";
 export { FederatedToken } from "./token";
+export { CompositeTokenSource, CookieTokenSource, HeaderTokenSource, type TokenSource } from "./tokensource";
 
-export { type TokenSource } from "./tokensource";
-export { CompositeTokenSource, CookieTokenSource, HeaderTokenSource, } from "./tokensource";
